requirements.txt
streamlit_gsheets==0.0.1
praat-parselmouth
pyaudio
scipy
IPython
playsound
os 
time
wave
sys
sounddevice
soundfile
numpy
pandas
datetime